+++ CHANGELOG.md
@@ -22,7 +22,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## 🚀 Features
 
-- **Add `--use-example-schema` flag to `subgraph publish`**
+- **Add `--use-example-schema` flag to `subgraph publish` - @samaanghani PR #3218**
 
   Allows publishing a placeholder schema without needing to provide your own schema file. This is useful for setting up your graph structure before your actual schemas are ready. The placeholder schema is `type Query { helloWorld: String }` with a routing URL of `https://example.com`.